Create screen to show events and show them

+++ app/src/main/java/br/com/osouza/projectdm114/order/OrderDetail.kt
@@ -2,7 +2,7 @@ package br.com.osouza.projectdm114.order
 
 data class OrderDetail(
     var username: String,
-    var orderId: Int,
+    var orderId: Long,
     var status: String,
     var productCode: String
 )
