+++ app/src/main/java/br/com/osouza/projectdm114/order/OrderDetailBindingAdapters.kt
@@ -4,7 +4,7 @@ import android.widget.TextView
 import androidx.databinding.BindingAdapter
 
 @BindingAdapter("orderId")
-fun bindOrderId(txtOrderId: TextView, orderId: Int?) {
+fun bindOrderId(txtOrderId: TextView, orderId: Long?) {
     txtOrderId?.let {
         txtOrderId.text = "$orderId"
     }
